# app/static/swagger.json
{
  "info": {
    "description": "Password management API created as a free software. Ready for deployment on server or PaaS like Heroku.",
    "version": "1.0.0",
    "title": "Open Password Management API",
    "contact": {
      "name": "Robert Molenda",
      "url": "https://github.com/Rojber"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "basePath": "/api",
  "swagger": "2.0",
  "securityDefinitions": {
    "UserToken": {
     "type": "apiKey",
     "in": "header",
     "name": "token"
    }
  },
  "definitions": {
...
    "LoginData_get": {
      "properties": {
        "_id": {
          "type": "object",
          "properties": {
            "$oid": {
              "type": "string"
            }
          }
        },
        "site": {
          "type": "string"
        },
        "login": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "passwordStrength": {
          "type": "integer"
        },
        "note": {
          "type": "string"
        }
      }
    },
    "LoginData_put": {
      "properties": {
        "site": {
          "type": "string"
        },
        "login": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "note": {
          "type": "string"
        }
      }
    },
...
  },
  "responses":{
    "UnauthorizedError":{
      "description": "{\"response\": \"WRONG TOKEN\"} or {\"response\": \"SESSION EXPIRED\"}"
    }
  },
  "paths": {
...
    "/LoginData/{loginID}": {
      "get": {
        "security": [
          {
            "UserToken": []
          }
        ],
        "summary": "Download JSON with full LoginData record of specified loginID",
        "parameters": [
          {
            "name": "loginID",
            "in": "path",
            "description": "ID of LoginData to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Returning user LoginData",
            "schema": {
              "$ref": "#/definitions/LoginData_get"
            }
          },
          "401": {
            "$ref": "#/responses/UnauthorizedError"
          }
        }
      },
      "put": {
        "security": [
          {
            "UserToken": []
          }
        ],
        "summary": "Edit LoginData record of specified loginID",
        "parameters": [
          {
            "name": "loginID",
            "in": "path",
            "description": "ID of LoginData to edit",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "LoginData to edit",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginData_put"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "{\"response\": \"OK\"}"
          },
          "401": {
            "$ref": "#/responses/UnauthorizedError"
          }
        }
      },
      "delete": {
        "security": [
          {
            "UserToken": []
          }
        ],
        "summary": "Delete LoginData record of specified loginID",
        "parameters": [
          {
            "name": "loginID",
            "in": "path",
            "description": "ID of LoginData to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "{\"response\": \"OK\"}"
          },
          "401": {
            "$ref": "#/responses/UnauthorizedError"
          }
        }
      }
    },
    "/LoginData": {
      "post": {
        "security": [
          {
            "UserToken": []
          }
        ],
        "summary": "Add LoginData record",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "LoginData to add",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginData_put"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "{\"response\": \"OK\"}"
          },
          "401": {
            "$ref": "#/responses/UnauthorizedError"
          }
        }
      }
    },
...
  }
}